+++ build-logic/src/test/kotlin/jp/juggler/konaResource/buildlogic/KonaBuildHostTest.kt
@@ -11,7 +11,7 @@ class KonaBuildHostTest : FreeSpec() {
             assertEquals(KonaBuildHost.LinuxX64, KonaBuildHost.from("Linux", "x86_64"))
             assertEquals(KonaBuildHost.MacosArm64, KonaBuildHost.from("Mac OS X", "aarch64"))
             assertEquals(KonaBuildHost.MacosX64, KonaBuildHost.from("Darwin", "amd64"))
-            assertEquals(KonaBuildHost.WindowsX64, KonaBuildHost.from("Windows 11", "x64"))
+            assertEquals(KonaBuildHost.MingwX64, KonaBuildHost.from("Windows 11", "x64"))
         }
 
         "rejectsUnsupportedHosts" {